tarea_De_movimiento/requirements.txt
pyserial>=3.5
numpy>=1.21.0
pandas>=1.3.0
tensorflow>=2.10.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
